Aggiornati messaggi eccezioni; completata pulizia codice; aggiunti Javadoc; aggiornate dipendenze; uniformate alcune chiamate a funzione

+++ src/main/java/it/disi/unitn/TracksMerger.java
@@ -1,5 +1,6 @@
 package it.disi.unitn;
 
+import it.disi.unitn.exceptions.InvalidArgumentException;
 import it.disi.unitn.exceptions.NotEnoughArgumentsException;
 import org.apache.commons.io.FileUtils;
 import org.jetbrains.annotations.NotNull;
@@ -22,8 +23,11 @@ public class TracksMerger {
 
     TracksMerger(@NotNull FFMpegBuilder builder, @NotNull String outputFile, @NotNull String audioInput,
                         @NotNull String videoInput) throws NotEnoughArgumentsException {
-        if(builder == null || audioInput == null || videoInput == null) {
-            throw new NotEnoughArgumentsException("The arguments to TracksMerger's constructor cannot be null.");
+        if(builder == null || audioInput == null || audioInput.isEmpty() || videoInput == null || videoInput.isEmpty() ||
+        outputFile == null || outputFile.isEmpty()) {
+            throw new NotEnoughArgumentsException("The arguments to this class's constructor cannot be null or empty " +
+                    "strings.", "Nessuno degli argomenti forniti al costruttore di questa classe puo' essere null o una " +
+                    "stringa vuota.");
         }
 
         this.builder = builder;
@@ -33,8 +37,10 @@ public class TracksMerger {
     }
 
     TracksMerger(@NotNull FFMpegBuilder builder, @NotNull String outputVideo) throws NotEnoughArgumentsException {
-        if(builder == null || outputVideo == null) {
-            throw new NotEnoughArgumentsException("The arguments to TracksMerger's constructor cannot be null.");
+        if(builder == null || outputVideo == null || outputVideo.isEmpty()) {
+            throw new NotEnoughArgumentsException("The arguments to this class's constructor cannot be null or empty " +
+                    "strings.", "Nessuno degli argomenti forniti al costruttore di questa classe puo' essere null o una " +
+                    "stringa vuota.");
         }
 
         this.builder = builder;
@@ -77,19 +83,21 @@ public class TracksMerger {
      * @param tempFile The path to the file that wil contain the paths of the files to be merged
      * @return The File instance of the file containing the paths that were passed to this method
      * @throws IOException if an I/O error occurred
+     * @throws InvalidArgumentException If any argument to this method is null or an empty string or contains null or
+     * empty strings
      */
-    private @NotNull File writeTXTFile(@NotNull List<String> inputFiles, @NotNull String tempFile) throws IOException {
-        if(inputFiles == null || tempFile == null) {
-            throw new IllegalArgumentException("No arguments to this method can be null.");
+    private @NotNull File writeTXTFile(@NotNull List<String> inputFiles, @NotNull String tempFile) throws IOException,
+            InvalidArgumentException {
+        boolean match = inputFiles.stream().anyMatch(s -> s == null || s.isEmpty());
+        if(inputFiles == null || match || tempFile == null || tempFile.isEmpty()) {
+            throw new InvalidArgumentException("No argument to this method can be null or an empty string, nor can it " +
+                    "contain null or empty strings.", "Nessuno degli argomenti forniti a questo metodo puo' essere null " +
+                    "o una stringa vuota o contenere stringhe null o vuote.");
         }
 
-        boolean match = inputFiles.stream().anyMatch(s -> s == null || s.isEmpty());
         File file = new File(tempFile);
         if(file.exists() || match) {
             file.delete();
-        }
-        if(match) {
-            throw new IllegalArgumentException("No arguments to this method can be null or empty strings.");
         }
         boolean created = file.createNewFile();
         for(String s: inputFiles) {
@@ -107,9 +115,18 @@ public class TracksMerger {
      * @param timeUnit The TimeUnit instance to be used
      * @param tempFile A temporary file used to store the paths of the files to be merged
      * @throws IOException if an I/O error occurs
+     * @throws InvalidArgumentException If any of the arguments given to this method is null, an empty string, is less
+     * than or equal to zero, or it contains null or empty strings
      */
     public void mergeVideos(long time, @NotNull TimeUnit timeUnit, @NotNull List<String> inputFiles,
-                            @NotNull String tempFile) throws IOException {
+                            @NotNull String tempFile) throws IOException, InvalidArgumentException {
+        if(time <= 0 || timeUnit == null || inputFiles == null || inputFiles.stream().anyMatch(s -> s == null || s.isEmpty()) ||
+        tempFile == null || tempFile.isEmpty()) {
+            throw new InvalidArgumentException("No argument to this method can be null, less than or equal to zero or " +
+                    " an empty string, nor can it contain null or empty strings.", "Nessuno degli argomenti forniti a " +
+                    "questo metodo puo' essere null, minore o uguale a zero o una stringa vuota o contenere stringhe null " +
+                    "o vuote.");
+        }
         File inputTXTFile = writeTXTFile(inputFiles, tempFile);
         builder.setCommand(builder.getCommand() + " -f concat -safe 0 -i \"" +
                 inputTXTFile.getPath().replace('\\', '/') + "\"");
